pulling out sandput functions so they can be generalised for all pipes, changing the output table so don't mess with nexisting reporting - not, think the sandput upload isn't working
